% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synonymous_snps_helpers.R
\name{analyze_transcript_integrity}
\alias{analyze_transcript_integrity}
\title{Pre-compute the Integrity State of Transcripts Affected by Variants}
\usage{
analyze_transcript_integrity(variants_genomic_on_ts, cds_by_tx, cds_seqs, txdb)
}
\description{
Pre-compute the Integrity State of Transcripts Affected by Variants
}
\keyword{internal}
