% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synonymous_snps_helpers.R
\name{annotate_mutations_with_alphagenome}
\alias{annotate_mutations_with_alphagenome}
\title{Annotate Variants with AlphaGenome Scores}
\usage{
annotate_mutations_with_alphagenome(
  all_variants,
  alphagenome_key,
  species,
  python_exec = "python3",
  alphagenome_context = ""
)
}
\arguments{
\item{all_variants}{A \code{GRanges} object. Must contain metadata columns
named \code{REF} and \code{ALT} for the reference and alternate alleles.}

\item{alphagenome_key}{Your key to the alphagenome service.}

\item{species}{A character string, either "human", "mouse".}

\item{python_exec}{path to the python3 that has alphagenome installed.}

\item{alphagenome_context}{Celltypes for which to filter alphagenome scores.}
}
\value{
alphagenome table of results, sorted as original \code{all_variants}
}
\description{
This function takes a GRanges object containing genetic variants, calls a
Python script to score them using the AlphaGenome API, and returns the
the new scores as DataFrameList.
}
\keyword{internal}
