% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synonymous_snps_helpers.R
\name{annotate_variants_with_cds}
\alias{annotate_variants_with_cds}
\title{Annotate SNV effects on CDS, considering background variants.}
\usage{
annotate_variants_with_cds(
  all_variants,
  variants_genomic_on_ts,
  txdb,
  genome,
  benign_if_moot_statuses = c("Frameshifted", "Splice site disrupted",
    "DOWNSTREAM_OF_PTC")
)
}
\arguments{
\item{all_variants}{A GRanges object of SNVs to be annotated.}

\item{variants_genomic_on_ts}{A GRanges object of background variants (SNVs or indels)
that define the "mutated" reference transcriptome.}

\item{txdb}{A TxDb object for transcript information.}

\item{genome}{A BSgenome object for sequence extraction.}

\item{benign_if_moot_statuses}{"Frameshifted", "Splice site disrupted", "DOWNSTREAM_OF_PTC"}
}
\value{
A DataFrameList, where each element corresponds to a variant in
\code{all_variants} and contains a DataFrame of annotations for each affected transcript.
}
\description{
This function annotates a set of single nucleotide variants (\code{all_variants})
by determining their effect on the coding sequences of overlapping transcripts.
It uniquely accounts for a pre-defined set of background variants
(\code{variants_genomic_on_ts}), which can include indels, that modify the
transcript sequences before the SNVs of interest are evaluated.
}
