% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synonymous_snps_helpers.R
\name{annotate_variants_with_dbsnp}
\alias{annotate_variants_with_dbsnp}
\title{Annotate variants with dbSNP data}
\usage{
annotate_variants_with_dbsnp(all_variants, snps, alt_allele_col = "ALT")
}
\description{
Annotate variants with dbSNP data
}
\keyword{internal}
