% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synonymous_snps_helpers.R
\name{annotate_variants_with_noncoding}
\alias{annotate_variants_with_noncoding}
\title{Check if our variants overlap something non-coding}
\usage{
annotate_variants_with_noncoding(all_variants, annotation)
}
\description{
Check if our variants overlap something non-coding
}
\keyword{internal}
