% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synonymous_snps_helpers.R
\name{annotate_variants_with_snps}
\alias{annotate_variants_with_snps}
\title{Check if our variants are knoww synonymous variants}
\usage{
annotate_variants_with_snps(all_variants, snps, alt_allele_col = "ALT")
}
\description{
Check if our variants are knoww synonymous variants
}
\keyword{internal}
