% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_mapping.R
\name{build_variant_layout}
\alias{build_variant_layout}
\title{Inject multiple variants and create a coordinate map (Vectorized)}
\usage{
build_variant_layout(variants_in_window, seq_len)
}
\arguments{
\item{variants_in_window}{A GRanges object of variants.
Must have names. Must be sorted.}

\item{seq_len}{A length of the sequence variants are mapped to.}
}
\value{
A GRanges object mapping the mutated coordinates.
}
\description{
This function takes a genomic sequence and a set of variants, injects the
ALT alleles, and produces a map to trace any position in the new mutated
sequence back to its origin (either the original genome or a specific variant).
This version is vectorized for performance and readability.
}
\keyword{internal}
