% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_helpers.R
\name{calculate_compatibility_map}
\alias{calculate_compatibility_map}
\title{Calculate compatibility map based on dbSNP information}
\usage{
calculate_compatibility_map(mutations)
}
\description{
Calculate compatibility map based on dbSNP information
}
\keyword{internal}
