% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants.R
\name{correct_variants}
\alias{correct_variants}
\title{Correct non-VCF compliant variants by adding genomic anchors.}
\usage{
correct_variants(variants, genome)
}
\arguments{
\item{variants}{A GRanges object of variants to check and correct.}

\item{genome}{A BSgenome object or similar for use with getSeq.}
}
\value{
A GRanges object where all validatable variants are now VCF-compliant.
}
\description{
This function uses \code{is_vcf_invalid} to identify variants that are non-compliant
due to empty alleles (e.g., REF="A", ALT=""). It corrects them by prepending
the preceding genomic base to both REF and ALT and adjusting the coordinates.
}
