% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_helpers.R
\name{create_template_and_probes}
\alias{create_template_and_probes}
\title{Create HDR template and probes from a set of mutations}
\usage{
create_template_and_probes(
  selected_muts,
  variants_in_editw,
  variants_genomic_on_ts,
  design_id,
  edit_region,
  source_genomic_seq,
  current_guide,
  do_probes,
  probe_params
)
}
\value{
A list containing the 'template' and 'probes' GRanges.
}
\description{
Injects selected SNPs into the base template sequence, calculates
summary statistics based on re-alignment, and designs validation probes.
}
\keyword{internal}
