% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_HDR_design_by_template.R
\name{design_by_template}
\alias{design_by_template}
\title{Design HDR probes and CRISPR guides}
\usage{
design_by_template(
  ensemble_transcript_id,
  mutation_loci,
  mutation_original,
  mutation_replacement,
  mutation_name,
  output_dir,
  annotation,
  genome = BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38,
  guide_distance = 10 + 17,
  extension = 400,
  positions_to_mutate = -19:19,
  mutations_per_template = 3,
  seed = 42
)
}
\arguments{
\item{ensemble_transcript_id}{This has to be ensemble transcript id e.g. ENST00000307851.9}

\item{mutation_loci}{Position on the gene that is mutated e.g. 245}

\item{mutation_original}{What was the original base on the genome? e.g. A}

\item{mutation_replacement}{What is the mutated base? e.g. G}

\item{mutation_name}{What is your name for this mutation e.g. tyr82cys}

\item{output_dir}{Where the files should be generated? Make sure you have writing permissions there.}

\item{annotation}{File path to the annotation file, a gff3.}

\item{genome}{BSgenome of your genome, compatible with annotation file, by default it is hg38.}

\item{guide_distance}{Window around which we should search for guides, relatively to the mutation loci. Defualt is 10 + 17bp.}

\item{extension}{How many bases upstream/downstream from the mutation loci should we include Default is 400.}

\item{positions_to_mutate}{Which positions from the mutation are available for mutation. By default its -30:30
leaving 20bp on each side of the template for the in case of incomplete integration of the template. Also codon
occupied by the mutation is not included in the change.}

\item{mutations_per_template}{How many codons should be mutated per template.}

\item{seed}{Ensures reproducibility of the random parts of the pipeline.}
}
\value{
writes files to the specified directory, might overwrite
}
\description{
Design HDR probes, CRISPR guides and NHEJ probes for a given
template transcript.
}
