% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_hdr.R
\name{design_hdr_internal}
\alias{design_hdr_internal}
\title{Internal workhorse function for the HDR design pipeline.}
\usage{
design_hdr_internal(params)
}
\arguments{
\item{params}{A list of all parameters passed from the user-facing function.}
}
\description{
Internal workhorse function for the HDR design pipeline.
}
\keyword{internal}
