% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_template.R
\name{design_one_template_for_all_guides}
\alias{design_one_template_for_all_guides}
\title{Design best CRISPR guide for fixing mutation}
\usage{
design_one_template_for_all_guides(
  ensemble_transcript_id,
  mutation_loci,
  mutation_original,
  mutation_replacement,
  mutation_name,
  output_dir,
  annotation,
  genome = BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38,
  guide_distance = 10 + 17,
  extension = 400,
  positions_to_mutate = -19:19,
  mutations_per_template = 3,
  seed = 42,
  intron_bp = 6,
  exon_bp = 3,
  snps = NULL,
  clinvar = NULL,
  score_efficiency = FALSE,
  probes = TRUE,
  primer3 = ""
)
}
\description{
Designs one optimized template for all of the guides overlapping site of interest.
Templates are optimized, by selecting synonymous SNPs that have the highest chance of disrupting
guides, minimizing disruption to other genomic functions by checking with submitted annotation file,
prioritizing SNPs that are already available in the population submitted through \verb{snps parameter. Guides are ranked first, by their inhibition potential, then by the quality of their snps, and further by efficiency scoring. The difference here to the }design_one_template_for_each_guide`
is in the fact that SNPs are optimized globally for all guides at the same time, because there will be
only one template for all of the guides.
}
