% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_helpers.R
\name{design_primers}
\alias{design_primers}
\title{Primer 3}
\usage{
design_primers(
  guide_distance,
  positions_to_mutate,
  genomic_seq,
  origin_mutation,
  mut_genomic,
  genome,
  primer3_path
)
}
\description{
Design primers with the use of primer 3
}
\keyword{internal}
