% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twisted_cas13a.R
\name{design_primers_rna}
\alias{design_primers_rna}
\title{Primer 3}
\usage{
design_primers_rna(
  target,
  position,
  guides_seq,
  primers_per_guide,
  primer3_path
)
}
\description{
Design primers with the use of primer 3 -
We want at least one primer to be outside of the template_range
We use \code{template_range_extended} as the region where we can design
We check for off-targets
We want to make sure the target_on_template_extended is captured inside the PCRed sequence
}
\keyword{internal}
