% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_template.R
\name{design_template}
\alias{design_template}
\title{Main function for the design of the templates.}
\usage{
design_template(one_for_all = FALSE)
}
\arguments{
\item{one_for_all}{One template for all of the guides.}
}
\description{
Main function for the design of the templates.
}
