% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twisted_cas13a.R
\name{design_twisted_cas13a}
\alias{design_twisted_cas13a}
\title{Design twisted Cas13a Guide Self-Folding}
\usage{
design_twisted_cas13a(
  input_fasta,
  position,
  guide_length = 28,
  allowed_positions = 3:5,
  t7_promoter = "GAAATTAATACGACTCACTATAGGG",
  linker = "GCGCT",
  primer3_path = Sys.which("primer3_core"),
  primers_per_guide = 2,
  rnaup_path = Sys.which("RNAup"),
  rnafold_path = Sys.which("RNAfold")
)
}
\arguments{
\item{input_fasta}{A DNAStringSet fasta file path.}

\item{position}{Position on input target to design for.}

\item{guide_length}{Normally 28}

\item{allowed_positions}{3:5, where we want guide to have mutation}

\item{t7_promoter}{"GAAATTAATACGACTCACTATAGGG" we attach this to left primer}

\item{linker}{"GCGCT" linker sequence - this is default optimized 5-mer}

\item{primer3_path}{Path to primer3_core.}

\item{primers_per_guide}{2, by default we design two primer pairs per guide}

\item{rnaup_path}{Path to RNAup binary.}

\item{rnafold_path}{Path to RNAfold binary.}
}
\value{
A data.frame with MFE (Minimum Free Energy) and structure notation.
}
\description{
Input: Patient RNA (SARS-CoV-2) or Genomic DNA (KRAS).
Conversion/Amplification:
If RNA: It is Reverse Transcribed into cDNA.
PCR Step: Primers amplify this DNA/cDNA.
Transcription (The Key Step): The PCR product is used as a template for T7 RNA Polymerase to create new RNA.
Detection: Cas13 detects this newly transcribed RNA.
Input defaults for tests
}
