% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_helpers.R
\name{export_design_results}
\alias{export_design_results}
\title{Export Design Results}
\usage{
export_design_results(
  output_dir,
  design_name,
  variant_genomic,
  var_data,
  guides,
  repair_template,
  genome,
  txdb,
  primers = GRanges(),
  probes_ = GRanges(),
  score_efficiency = FALSE,
  one_for_all = FALSE
)
}
\description{
Writes all final output files (FASTA, CSV, GFF3, GenBank).
}
\keyword{internal}
