% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_helpers.R
\name{find_best_snps_for_guide}
\alias{find_best_snps_for_guide}
\title{Find the best set of SNPs for a single guide}
\usage{
find_best_snps_for_guide(guide_snps, N, optimization_scheme)
}
\value{
A GRanges object of the selected mutations.
}
\description{
Sorts candidate SNPs based on a chosen scheme and greedily
selects the top N non-overlapping ones.
}
\keyword{internal}
