% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_helpers.R
\name{find_best_snvs_for_guide}
\alias{find_best_snvs_for_guide}
\title{Find the best set of SNVs for a single guide}
\usage{
find_best_snvs_for_guide(guide_snvs, N, optimization_scheme)
}
\value{
A GRanges object of the selected variants.
}
\description{
Sorts candidate SNVs based on a chosen scheme and greedily
selects the top N non-overlapping ones.
}
\keyword{internal}
