% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_helpers.R
\name{format_output_muts}
\alias{format_output_muts}
\title{Create the final summary list output}
\usage{
format_output_muts(selected_muts, guides, pams, strategy)
}
\arguments{
\item{selected_muts}{The final GRanges of selected mutations.}

\item{guides}{The GRanges of guides (can be single or multiple).}

\item{pams}{The GRanges of pams (can be single or multiple).}

\item{strategy}{The selection strategy used.}
}
\value{
A list with the standard output format.
}
\description{
Create the final summary list output
}
\keyword{internal}
