% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_combinations_of_mutations_for_guides}
\alias{get_combinations_of_mutations_for_guides}
\title{All combinations of mutations}
\usage{
get_combinations_of_mutations_for_guides(
  mutations,
  mutations_per_template,
  pam,
  guide
)
}
\arguments{
\item{mutations}{all possible mutations}

\item{mutations_per_template}{How many to select?}

\item{pam}{PAM of the guide}

\item{guide}{guide}
}
\value{
A list of muts grouped
}
\description{
this is more global version of above
we want to find the SNPs that disable as many of the guides as we can
we design for single template for all guides
}
