% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdna_diversification.R
\name{get_genomic_seq}
\alias{get_genomic_seq}
\title{Get genomic seqeunce as a fasta file}
\usage{
get_genomic_seq(
  chromosome = "chr1",
  position = 2313,
  bp = 3000,
  homology_arm_length = 400,
  output_file = NULL,
  genome = BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38
)
}
\arguments{
\item{chromosome}{What is the chromosome of your target site?}

\item{position}{Position on the chromosome}

\item{bp}{How many bp of the seqeunce around the cut site you would like to extract?}

\item{homology_arm_length}{How long should homology arms be?}

\item{output_file}{To which file output the results? Writes fasta file. By default it does not save the file.}

\item{genome}{Which genome to use? By default it uses BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38.}
}
\value{
returns the DNAStringSet of the extracted sequences
}
\description{
Outputs genomic sequence and homology arms of the specified lengths
}
