% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guides_and_scores_helpers.R
\name{get_guides_and_scores}
\alias{get_guides_and_scores}
\title{Get and Score Guides (Refactored)}
\usage{
get_guides_and_scores(
  variants_genomic,
  design_name,
  cut_distance_max,
  genome,
  ALT_on_genome,
  score_efficiency = TRUE
)
}
\arguments{
\item{variants_genomic}{A GRanges object of representing the variants.
Must have an 'ALT' metadata column.}

\item{design_name}{A character string for naming the design.}

\item{cut_distance_max}{An integer.}

\item{genome}{A BSgenome object for sequence retrieval.}

\item{ALT_on_genome}{whether any ALT variant is on the guides}

\item{score_efficiency}{A logical flag. If TRUE, calculate efficiency scores.}
}
\value{
A GRanges object containing all found guides, their sequences, scores
(for successfully run algorithms), and a final aggregated rank.
}
\description{
For a given genomic location, this function introduces a mutation,
designs potential CRISPR guides, and scores them for efficiency. It is designed
to be robust to failures in individual scoring algorithms.
}
\keyword{internal}
