% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twisted_cas13a.R
\name{get_offtarget_count_MM0}
\alias{get_offtarget_count_MM0}
\title{Check 0MM off-targets for each guide}
\usage{
get_offtarget_count_MM0(guides_seq, offtargets_seq)
}
\arguments{
\item{guides_seq}{DNAStringSet of the guides}

\item{offtargets_seq}{list of DNAStringSet elements where we will look for
0MM off-targets and sum up results}
}
\value{
A vector with one number for each guide
}
\description{
Check 0MM off-targets for each guide inside the list of
offtargets_seq.
}
\keyword{internal}
