% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDRdesignForCRISPR.R
\docType{package}
\name{HDR.design.for.CRISPR}
\alias{HDR.design.for.CRISPR-package}
\alias{HDR.design.for.CRISPR}
\title{HDR.design.for.CRISPR package}
\description{
Design HDR repair for CRISPR.
}
\author{
\strong{Maintainer}: Kornel Labun \email{email@kornellabun.com}

}
