% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synonymous_snps_helpers.R
\name{is_outside_cds_boundaries}
\alias{is_outside_cds_boundaries}
\title{Check if variants are outside CDS start/stop boundaries.}
\usage{
is_outside_cds_boundaries(all_variants, txdb, buffer_bp = 3)
}
\description{
Check if variants are outside CDS start/stop boundaries.
}
\keyword{internal}
