% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synonymous_snps_helpers.R
\name{is_outside_splice_sites}
\alias{is_outside_splice_sites}
\title{Check positions that are not over splice sites}
\usage{
is_outside_splice_sites(all_variants, txdb, intron_bp, exon_bp)
}
\description{
Check positions that are not over splice sites
}
\keyword{internal}
