% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants.R
\name{is_vcf_invalid}
\alias{is_vcf_invalid}
\title{Find variants that are not represented in a VCF-compliant manner.}
\usage{
is_vcf_invalid(variants)
}
\arguments{
\item{variants}{A GRanges object representing variants. Must have 'REF'
and 'ALT' metadata columns.}
}
\value{
An integer vector of indices corresponding to non-compliant variants.
}
\description{
This function checks for two common VCF representation rules:
\enumerate{
\item REF and ALT alleles must not be empty strings.
\item For indels (where nchar(REF) != nchar(ALT)), the first base of REF and ALT
must be identical.
}
}
