% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{melting_temp_vec}
\alias{melting_temp_vec}
\title{Vectorized Melting Temperature Calculation (Primer3Plus variant)}
\usage{
melting_temp_vec(x, Na = 50)
}
\arguments{
\item{x}{A DNAStringSet object.}

\item{Na}{The concentration of Na+ in mM.}
}
\value{
A numeric vector of melting temperatures.
}
\description{
This function calculates the melting temperature for a DNAStringSet using the
TmCalculator::Tm_GC Primer3Plus formula.
Primer3Plus: Tm = 81.5 + 0.41(Percentage_GC) - 600/N + 16.6 x log(Na+)
It is significantly faster than calling original Tm_GC.
}
