% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mirror_flip}
\alias{mirror_flip}
\title{Flip IRanges around a central mutation point.}
\usage{
mirror_flip(ranges, extension)
}
\arguments{
\item{ranges}{An IRanges object containing the ranges to be flipped, ranges in \code{sense}}

\item{extension}{An integer defining the "half-length" of the genomic_seq.
The total length will be \code{2 * extension + 1}, and the mirror point will be \code{extension + 1}.}
}
\value{
A new IRanges object with the ranges flipped.
}
\description{
Flip IRanges around a central mutation point.
}
