% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twisted_cas13a.R
\name{predict_spacer_amplicon_interaction}
\alias{predict_spacer_amplicon_interaction}
\title{Predict Cas13a Guide-Target Interaction using RNAup}
\usage{
predict_spacer_amplicon_interaction(
  spacer_with_linker_seq,
  amp_seq_rna,
  rnaup_path = "RNAup",
  window_size = 30
)
}
\arguments{
\item{spacer_with_linker_seq}{A RNAStringSet or character vector of spacer sequences with LINKER!}

\item{amp_seq_rna}{A RNAStringSet of amplicon sequences}

\item{rnaup_path}{Path to the RNAup binary.}

\item{window_size}{Integer. Max interaction length (roughly spacer length). Default 30.}
}
\value{
A data.frame containing:
\item{dG_binding}{Total interaction energy (lower is better)}
\item{dG_duplex}{Energy of hybridization}
\item{dG_opening}{Energy required to open the target structure}
}
\description{
Calculates the thermodynamic properties of the interaction
between Cas13a guides and PCR amplicons. It considers both the hybridization
energy and the accessibility of the target site within the amplicon.
}
\keyword{internal}
