% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twisted_cas13a.R
\name{predict_spacer_self_structure}
\alias{predict_spacer_self_structure}
\title{Check Cas13a Guide Self-Folding}
\usage{
predict_spacer_self_structure(spacer_with_linker_seq, rnafold_path = "RNAfold")
}
\arguments{
\item{spacer_with_linker_seq}{A DNAStringSet or character vector of spacer sequences.}

\item{rnafold_path}{Path to RNAfold binary.}
}
\value{
A data.frame with MFE (Minimum Free Energy) and structure notation.
}
\description{
Predicts the secondary structure of the full guide RNA
(Direct Repeat + Spacer). High internal structure in the spacer region
suggests poor performance.
}
\keyword{internal}
