% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_helpers.R
\name{prepare_candidate_snps}
\alias{prepare_candidate_snps}
\title{Prepare Candidate Synonymous Mutations}
\usage{
prepare_candidate_snps(
  env_data,
  positions_to_mutate,
  mutation_loci,
  mutation_name,
  intron_bp,
  exon_bp,
  clinvar,
  snps,
  cadd,
  annotation,
  genome
)
}
\description{
Generates, filters, and annotates all possible synonymous SNPs.
}
\keyword{internal}
