% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_mapping.R
\name{remap_target_to_genomic}
\alias{remap_target_to_genomic}
\title{Describe target composition in genomic and variant coordinates}
\usage{
remap_target_to_genomic(
  target,
  coordinate_map,
  window_genomic,
  variants_genomic
)
}
\arguments{
\item{target}{A GRanges object with ranges relative to the mutated sequence.}

\item{coordinate_map}{The map created by \code{map_variants}.}

\item{window_genomic}{The GRanges object for the entire genomic window.}

\item{variants_genomic}{The GRanges of the variants with genomic coordinates.}
}
\value{
A data.frame containing the original information for each target that
could be successfully mapped, with an added 'coords' metadata column.
The 'coords' column is a semicolon-separated string detailing the genomic
ranges (e.g., 'chr1:100-120') and variant parts (e.g., 'var_rs123:1-10')
that the target covers.
}
\description{
Using a coordinate map, this function translates the location of each target
on the mutated sequence into a detailed string describing its constituent
genomic and variant parts.
}
\keyword{internal}
