% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_mapping.R
\name{reverse_variants}
\alias{reverse_variants}
\title{Create a set of variants that can reverse a mutation injection process.}
\usage{
reverse_variants(variants)
}
\arguments{
\item{variants}{A GRanges object of the original variants. Must contain
'REF' and 'ALT' metadata columns. Object has to have names.}
}
\value{
A GRanges object representing the reverse variants. The ranges are in
the coordinate system of the \emph{mutated} sequence. The 'ALT' metadata column
contains the sequence needed to revert the mutation (i.e., the original 'REF').
}
\description{
Create a set of variants that can reverse a mutation injection process.
}
