% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_helpers.R
\name{select_non_overlapping_mutations}
\alias{select_non_overlapping_mutations}
\title{Selects mutations that are non-overlapping at both the genomic and codon level.}
\usage{
select_non_overlapping_mutations(gr, N)
}
\arguments{
\item{gr}{A GRanges object of candidate mutations, assumed to be pre-sorted by priority.}

\item{N}{The maximum number of mutations to select.}
}
\value{
A GRanges object containing the top N non-overlapping mutations.
}
\description{
Selects mutations that are non-overlapping at both the genomic and codon level.
}
\keyword{internal}
