% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_helpers.R
\name{select_non_overlapping_variants}
\alias{select_non_overlapping_variants}
\title{Selects variants that are non-overlapping at both the genomic and codon level.}
\usage{
select_non_overlapping_variants(gr, N)
}
\arguments{
\item{gr}{A GRanges object of candidate variants, assumed to be pre-sorted by priority.}

\item{N}{The maximum number of variants to select.}
}
\value{
A GRanges object containing the top N non-overlapping variants.
}
\description{
Selects variants that are non-overlapping at both the genomic and codon level.
}
\keyword{internal}
