% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_helpers.R
\name{setup_design_environment}
\alias{setup_design_environment}
\title{Setup Design Environment}
\usage{
setup_design_environment(
  ensemble_transcript_id,
  mutation_loci,
  mutation_original,
  annotation,
  genome
)
}
\description{
Loads and validates all initial transcript and genomic data.
}
\keyword{internal}
