% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdna_diversification.R
\name{synonymously_mutate_cds}
\alias{synonymously_mutate_cds}
\title{Design templates with innocent synonymous SNPs}
\usage{
synonymously_mutate_cds(
  ensemble_transcript_id,
  annotation,
  output_file,
  genome = BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38,
  seed = 42
)
}
\arguments{
\item{ensemble_transcript_id}{This has to be ensemble transcript id e.g. ENST00000307851.9}

\item{annotation}{File path to the annotation file, a gff3.}

\item{output_file}{Path and name of the out fasta file e.g "~/mutated_cds.fa"}

\item{genome}{BSgenome of your genome, compatible with annotation file, by default it is hg38.}

\item{seed}{Ensures reproducibility of the random parts of the pipeline.}
}
\value{
writes file to \code{output_file}, might overwrite
}
\description{
Design templates with mutations that will not change coding sequence
}
