% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_template.R
\name{synonymously_mutate_template}
\alias{synonymously_mutate_template}
\title{Design templates with innocent synonymous SNPs}
\usage{
synonymously_mutate_template(
  ensemble_transcript_id,
  mutation_loci,
  mutation_original,
  mutation_replacement,
  mutation_name,
  output_dir,
  annotation,
  genome = BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38,
  extension = 50,
  positions_to_mutate = -30:30,
  mutations_per_template = 4,
  n = 96,
  seed = 42
)
}
\arguments{
\item{ensemble_transcript_id}{This has to be ensemble transcript id e.g. ENST00000307851.9}

\item{mutation_loci}{Position on the gene that is mutated e.g. 245}

\item{mutation_original}{What was the original base on the genome? e.g. A}

\item{mutation_replacement}{What is the mutated base? e.g. G}

\item{mutation_name}{What is your name for this mutation e.g. tyr82cys}

\item{output_dir}{Where the files should be generated? Make sure you have writing permissions there.}

\item{annotation}{File path to the annotation file, a gff3.}

\item{genome}{BSgenome of your genome, compatible with annotation file, by default it is hg38.}

\item{extension}{How many bases upstream/downstream from the mutation loci should we include Default is 50.}

\item{positions_to_mutate}{Which positions from the mutation are available for mutation. By default its -30:30
leaving 20bp on each side of the template for the in case of incomplete integration of the template. Also codon
occupied by the mutation is not included in the change.}

\item{mutations_per_template}{How many codons should be mutated per template.}

\item{n}{How many tempaltes to generate? Defalt is set to 96.}

\item{seed}{Ensures reproducibility of the random parts of the pipeline.}
}
\value{
writes files to the specified directory, might overwrite
}
\description{
Design templates with mutations that will not change coding sequence
}
