% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synonymous_snps_helpers.R
\name{annotate_mutations_with_alphagenome}
\alias{annotate_mutations_with_alphagenome}
\title{Annotate Variants with AlphaGenome Scores}
\usage{
annotate_mutations_with_alphagenome(
  all_variants,
  alphagenome_key,
  species,
  python_exec = "python3",
  alphagenome_context = ""
)
}
\arguments{
\item{all_variants}{A \code{GRanges} object. Must contain metadata columns
named \code{REF} and \code{ALT} for the reference and alternate alleles.}

\item{alphagenome_key}{A character string containing your private AlphaGenome API key.}

\item{species}{A character string, either "human", "mouse".}

\item{python_exec}{The command to execute Python (e.g., "python", "python3").
Default is "python".}
}
\value{
\code{all_variants} with alphagenome mcols
}
\description{
This function takes a GRanges object containing genetic variants, calls a
Python script to score them using the AlphaGenome API, and returns the
the new scores as DataFrameList.
}
\keyword{internal}
