% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synonymous_snps_helpers.R
\name{annotate_variants_with_cds}
\alias{annotate_variants_with_cds}
\title{Annotate coding sequence changes for a set of variants.}
\usage{
annotate_variants_with_cds(all_variants, txdb, genome)
}
\arguments{
\item{all_variants}{A GRanges object containing the variants. Must have "REF"
and "ALT" metadata columns and be named.}

\item{txdb}{A TxDb object containing transcript models.}

\item{genome}{A BSgenome object for the reference genome.}
}
\value{
A DataFrameList where each element corresponds to a variant. Each
DataFrame contains detailed annotations for every transcript the variant
overlaps, with columns: tx_id, tx_strand, codon_ref, codon_alt,
aa_ref, aa_alt, codon_num, and frame.
}
\description{
This function takes a GRanges object of single nucleotide variants (SNVs),
all on the '+' strand, and annotates their effects on the coding sequences (CDS)
of overlapping transcripts.
}
\keyword{internal}
