% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_hdr.R
\name{create_template_and_probes}
\alias{create_template_and_probes}
\title{This new helper function abstracts the creation of a single template and its probes.
It will be used inside the switch statement to reduce code duplication.}
\usage{
create_template_and_probes(
  muts,
  guide_name,
  origin_mutation,
  template_range,
  genomic_seq,
  mutation_name,
  do_probes,
  probe_params
)
}
\description{
This new helper function abstracts the creation of a single template and its probes.
It will be used inside the switch statement to reduce code duplication.
}
\keyword{internal}
