% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_hdr.R
\name{design_hdr}
\alias{design_hdr}
\title{Design HDR Repair Templates for CRISPR}
\usage{
design_hdr(
  design_name,
  chrom,
  variant_start,
  variant_end,
  REF,
  ALT,
  ALT_on_genome,
  ALT_on_templates,
  output_dir,
  annotation,
  genome = BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38,
  optimization_scheme = "balanced",
  maximum_mutations_per_template = 3,
  filter_to_guide = "",
  cut_distance_max = 30,
  template_size = 120,
  template_hr_arm_size = 30,
  seed = 42,
  intron_bp = 6,
  exon_bp = 3,
  snps = NULL,
  clinvar = NULL,
  cadd = NULL,
  benign_cadd_threshold = 15,
  score_efficiency = FALSE,
  do_probes = TRUE,
  primer3 = "",
  alphagenome_key = "",
  alphagenome_context = "",
  alphagenome_threshold = 0.99,
  splicing_count = 1,
  python_exec = "python3"
)
}
\arguments{
\item{design_name}{A short name for this set of mutations, e.g., "tyr82cys".0,}

\item{chrom}{Chromosome of the variant to correct. e.g. chr1}

\item{variant_start}{Position on the chromosome for the variants (its on + strand)}

\item{variant_end}{Position ends of the variants on the chromosome}

\item{REF}{The original bases on the genome, e.g., "A".}

\item{ALT}{The desired mutated bases, e.g., "G".}

\item{ALT_on_genome}{A vector of T/F for each variant. Guides should match your target cell sequences. Select false when targeting wild type.}

\item{ALT_on_templates}{A vector of T/F for each variant. This should match your desired outcome. Select true when correcting alt variant.}

\item{output_dir}{Path to the directory where output files will be saved.}

\item{annotation}{File path to the genome annotation file (GFF3/GTF) or (.db/.sqlite) that can be loaded with \code{AnnotationDbi::loadDb}.}

\item{genome}{A \code{BSgenome} object for your genome, e.g., \code{BSgenome.Hsapiens.UCSC.hg38}.}

\item{optimization_scheme}{The optimization scheme for selecting synonymous SNPs. Must be one of:
\itemize{
\item \code{"balanced"}: (Default) Balances safety, PAM disruption, and SNP quality.
\item \code{"safety_first"}: Prioritizes avoiding non-coding overlaps and using known benign SNPs.
\item \code{"disruption_first"}: Prioritizes PAM disruption and guide disruption.
}}

\item{maximum_mutations_per_template}{The maximum number of synonymous SNPs to introduce per repair template.}

\item{filter_to_guide}{Optional. A 20bp character string of a specific guide sequence. If provided, the design process will be restricted to only this guide.}

\item{cut_distance_max}{Window around the variants to search for cut of the guides (default: 30).}

\item{template_size}{Size of the repair template (default: 120).}

\item{template_hr_arm_size}{Size of the repair template (default: 30).}

\item{seed}{An integer for setting the random seed to ensure reproducibility.}

\item{intron_bp}{Number of intronic bases near splice sites to exclude from synonymous SNPs (default: 6).}

\item{exon_bp}{Number of exonic bases near splice sites to exclude from synonymous SNPs (default: 3).}

\item{snps}{Optional. An object like \code{SNPlocs.Hsapiens.dbSNP155.GRCh38}.}

\item{clinvar}{Optional. File path to a ClinVar VCF file.}

\item{cadd}{Optional. A function to retrieve CADD scores, e.g., \code{getGScores("cadd.v1.6.hg38")}.}

\item{benign_cadd_threshold}{Numeric. CADD score below which a variant is
considered high-confidence benign. Default 15.}

\item{score_efficiency}{Logical. If \code{TRUE}, score guides using models from the \code{crisprScore} package.}

\item{do_probes}{Logical. If \code{TRUE}, design qPCR probes for HDR, NHEJ, and Reference.}

\item{primer3}{Optional. Full path to the \code{primer3_core} executable to design PCR primers.}

\item{alphagenome_key}{Your key to the alphagenome service.}

\item{alphagenome_context}{Celltypes for which to filter alphagenome scores.}

\item{alphagenome_threshold}{This is the threshold used to decide if a given splicing
prediction is not considered safe when above it}

\item{splicing_count}{This is the number of splicing predictors that have to be above the
\code{alphagenome_threshold} to consider a SNV as unsafe. Because we are using max
across all tissues and cell types we want to be sure an SNV is dangerous.}

\item{python_exec}{path to the python3 that has alphagenome installed.}
}
\value{
This function does not return a value. It writes all output files to the specified \code{output_dir}.
}
\description{
A comprehensive function to design HDR repair templates. It can generate templates
using different strategies: finding one optimal template for each guide, finding one optimal
template for all guides, or generating all possible template combinations.
}
