% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/same_template_all_guides.R
\name{design_one_template_for_all_guides}
\alias{design_one_template_for_all_guides}
\title{Design best CRISPR guide for fixing mutation}
\usage{
design_one_template_for_all_guides(
  ensemble_transcript_id,
  mutation_loci,
  mutation_original,
  mutation_replacement,
  mutation_name,
  output_dir,
  annotation,
  genome = BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38,
  guide_distance = 10 + 17,
  extension = 400,
  positions_to_mutate = -19:19,
  mutations_per_template = 3,
  seed = 42,
  intron_bp = 2,
  exon_bp = 0,
  snps = NULL,
  clinvar = NULL,
  score_efficiency = FALSE
)
}
\arguments{
\item{ensemble_transcript_id}{This has to be ensemble transcript id e.g. ENST00000307851.9}

\item{mutation_loci}{Position on the gene that is mutated e.g. 245}

\item{mutation_original}{What was the original base on the genome? e.g. A}

\item{mutation_replacement}{What is the mutated base? e.g. G}

\item{mutation_name}{What is your name for this mutation e.g. tyr82cys}

\item{output_dir}{Where the files should be generated? Make sure you have writing permissions there.}

\item{annotation}{File path to the annotation file, a gff3.}

\item{genome}{BSgenome of your genome, compatible with annotation file, by default it is hg38.}

\item{guide_distance}{Window around which we should search for guides, relatively to the mutation loci. Defualt is 10 + 17bp.}

\item{extension}{How many bases upstream/downstream from the mutation loci should we include. Default is 400.}

\item{positions_to_mutate}{Which positions from the mutation are available for mutation. By default its -30:30
leaving 20bp on each side of the template for the in case of incomplete integration of the template. Also codon
occupied by the mutation is not included in the change.}

\item{mutations_per_template}{How many codons should be mutated per template.}

\item{seed}{Ensures reproducibility of the random parts of the pipeline.}

\item{intron_bp}{How many bp of the intronic part of splicing should be excluded from synonymous SNPs (default 2)?}

\item{exon_bp}{How many bp of the exonic part of splicing should be excluded from synonymous SNPs (default 0)?}

\item{snps}{Can be either NULL or an object like SNPlocs.Hsapiens.dbSNP155.GRCh38 from library(SNPlocs.Hsapiens.dbSNP155.GRCh38)}

\item{clinvar}{This is a clinVar database VCF file location. You can download it here https://ftp.ncbi.nlm.nih.gov/pub/clinvar/vcf_GRCh38/clinvar.vcf.gz and https://ftp.ncbi.nlm.nih.gov/pub/clinvar/vcf_GRCh38/clinvar.vcf.gz.tbi}

\item{score_efficiency}{If you have installed relevant models from crisprScore package, you can set it to true to include scores from these models. Default is FALSE.}
}
\value{
writes files to the specified directory, might overwrite
}
\description{
Designs one optimized template for all of the guides overlapping site of interest.
Templates are optimized, by selecting synonymous SNPs that have the highest chance of disrupting
guides, minimizing disruption to other genomic functions by checking with submitted annotation file,
prioritizing SNPs that are already available in the population submitted through \verb{snps parameter. Guides are ranked first, by their inhibition potential, then by the quality of their snps, and further by efficiency scoring. The difference here to the }design_one_template_for_each_guide`
is in the fact that SNPs are optimized globally for all guides at the same time, because there will be
only one template for all of the guides.
}
