% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synonymous_snps_helpers.R
\name{prepare_candidate_snps}
\alias{prepare_candidate_snps}
\title{Prepare Candidate Synonymous Mutations}
\usage{
prepare_candidate_snps(
  annotation,
  txdb,
  genome,
  variant_genomic,
  allowed_positions_upstream,
  allowed_positions_downstream,
  intron_bp,
  exon_bp,
  clinvar,
  snps,
  cadd,
  alphagenome_key,
  python_exec
)
}
\description{
Generates, filters, and annotates all possible synonymous SNPs.
}
\keyword{internal}
