% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synonymous_snps_helpers.R
\name{prepare_candidate_snps}
\alias{prepare_candidate_snps}
\title{Prepare Candidate Synonymous Mutations}
\usage{
prepare_candidate_snps(
  candidate_snp_map,
  annotation,
  txdb,
  genome,
  variants_genomic_on_ts,
  intron_bp,
  exon_bp,
  clinvar,
  snps,
  cadd,
  alphagenome_key,
  python_exec,
  alphagenome_context
)
}
\description{
Generates, filters, and annotates all possible synonymous SNPs.
}
\keyword{internal}
