% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guides_and_scores_helpers.R
\name{process_strand}
\alias{process_strand}
\title{Helper function to find, score, and format guides for one strand}
\usage{
process_strand(
  mutated_seq,
  window_genomic,
  pam_pattern,
  strand_char,
  scorers,
  score_efficiency
)
}
\description{
Helper function to find, score, and format guides for one strand
}
\keyword{internal}
