% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_helpers.R
\name{write_component_files}
\alias{write_component_files}
\title{Helper write components}
\usage{
write_component_files(
  output_dir,
  design_name,
  component,
  component_name,
  chrom_relative
)
}
\description{
Helper to write CSV and GFF3 for a given GRanges component.
It calculates relative coordinates within the \code{chrom_relative} window.
}
\keyword{internal}
